Add check for Game1.currentLocation null to OnUpdateTicked.
This eliminates a NullReferenceException printed to the log when
changing locations.  (I only saw this as a farm hand.  My specific
test case was to repeatedly enter the greenhouse in SVE.  It was
also raining, but I'm not sure that had anything to do with it.
The error appeared most (but not all) of the time.)