Improve performance with other sprinkler mods; v2.5.0
Improved performance of highlighting when integrated with other sprinkler
mods by shifting some per-item calculations to per-batch.  (I.e., slightly
increased the fixed cost overhead of OnUpdateTicked in order to decrease
the per-item cost.)

Add a new signature of `AddItemRangeHighlighter` with callbacks for the
beginning and end of each "batch" of highlight range calculation.