hl at action loc for held items, don't hl at cursor if mouse hidden
When enabled, if the currently held item has a highlighter then show the
range at the location where the item would be placed if you press the
action button.  _Don't_ show the range at the mouse location if the mouse
is hidden (which seems to never be the case with the controller I have,
but maybe is with fancier controllers).