fix bug in Better Junimos integration; version 3.6.1
Previously, the junimo hut range was fixed when the integration was
installed, which is annoying when it means that in-game changes to the
configuration are not reflected until returning to the title screen, but a
serious problem when using the "Junimos work for wages" option in Better
Junimos.  When using that option, the range is initially a reduced range
until the junimos have been paid.

The range is now re-fetched before returning the highlight tiles for
junimo huts.